Implemented the first spreadsheet: Mining
This is the most basic one, it takes all the ore variants in the database and calculates the revenue by Km^3 based on the best buy price for each ore variant, the prices come from http://eve-central.com/.
It's coded to allow the user to specify the system's prices for the calculation, but right now it just defaults to the Amarr system.
The coding also allows to take sale taxes into account, but right now it defaults to 0% sale taxes.