Added two methods to check EVE-Central IDs
One is for Mineral and the other is for Variation, what they do is send a
request to eve-central using the ID stored in the local database, and
compare the item name returned by eve-central with the stored item name.

If both names are the same then the eve-central ID is correct for said
item. Currently only works for Minerals and Ore Variations since those
are the only items which we use for price checks at eve-central.